Conflicts:
	Sources/Internal/Platform/Qt5/MacOS/OpenGLView.mm
	Sources/Internal/Platform/Qt5/QtLayer.cpp
	Sources/Internal/Platform/Qt5/Win32/QtLayerWin32.h
	Sources/Internal/Render/RenderManager.cpp
	Sources/Internal/Render/RenderManager.h
	Tools/ResourceEditor/Classes/Qt/DockSceneTree/SceneTreeItem.cpp
	Tools/ResourceEditor/Classes/Qt/Main/davaglwidget.cpp
	Tools/ResourceEditor/Classes/Qt/Main/davaglwidget.h

Former-commit-id: ab2f9fb9977938fb267863fb4499d87481bfb991

+++ Sources/Internal/Render/2D/GraphicsFont.cpp
@@ -358,7 +358,7 @@ Font::StringMetrics GraphicsFont::GetStringMetrics(const WideString& str, Vector
 
 Font::StringMetrics GraphicsFont::DrawString(float32 x, float32 y, const WideString & string, int32 justifyWidth, int32 spaceAddon, Vector<float32> *charSizes, bool draw) const
 {
-	const uint32 length = string.length();
+	const uint32 length = static_cast<uint32>(string.length());
 	if(length == 0) return Font::StringMetrics();
 
 	uint32 countSpace = 0;
